Prevent property sheets with conflicting fields.
We now enforce unique fields names within certain subsets of slots.

This has become necessary as our Solr dynamic field implementation
potentially shares field names between different property sheets. To
prevent undefined behavior when a sheet for the default slot and a sheet
for atype specific slot define the same field, we don't allow this to
happen in the first place.

We also don't allow a property sheet to be assigned to the default slot
and a type specific slot at the same time.

The implementation here is a bit more strict that would be necessary
as same name but different type would currently result in different
Solr dynamic field names. I think it is a reasonable and future proof
restrictions, and users can just choose a new name like `title_1` to
avoid conflicts.