Scan uploaded file for every file field.
Scan results get stored on the request to avoid scanning
a given file several times in the same request. Because the key used
to store the scan results did not depend on the field, the scan
result of the first scanned field would be used for all other
file fields with uploaded files in the same request. We fix this
by making the key specific for each field.