Download view does not swallow exceptions by default.
We cannot rely on the request headers to find out whether a request
is made over the API to the download view. Therefore the view
should not swallow exceptions when they happen but raise them by
default. To display a status message and redirect in the classic UI
we now pass an additional parameter error_as_message, which will
display exceptions as status messages instead.