Support solr stats component in `@solrsearch`.
Add support to return simple statistics for numeric, string and date
fields as specified in the solr documentation at
https://solr.apache.org/guide/8_8/the-stats-component.html.

As with the facet endpoint the query is passed more or less directly to
solr. To receive the stats component in the response to solrsearch
both enabling it with `stats=true` and passing all the the desired
fields by repeating the `stats.field=<fieldname>` parameter is
necessary.
We take care to not return any statistics for the blacklisted fields.