Remove strict argument for migrations.
The argument was somehow implemented by us with the initial migrations
and the re-implemented in all new migrations. But it cannot be used when
running migrations via browser and we do not seem to set it to False
anywhere but in tests. Also the IPreMigrationHook does not officially
provide and support this argument.

I've opted to remove the unusable argument and to simplify the current
implementation.