Add tests for trasher.
Before refactoring the trash module, we add tests for the different
portal types. Note that the ITrashableMarker behavior is provided
for diverse portal types which cannot be trashed. For those, trashing
will raise Unauthorized if the "Trash content" permission is not attributed
on that content type, or an AttributeError if it is, because verify_may_trash
will fail.