Fix sending additional files to pdflatex service.
Fix two bugs with the pdflatex service when sending additional files:
- Instead of the file content we would just send the path to the file
- We would not send the filename, leading to every file be prefixed with
  `.file` upon save. This broke references to the file.

I have opted to read the file myself before passing it to requests to be
in control of closing the file once it is read. Requests does not take
care of this and reading the files in a context manager guarantees that
we are in control of closing. I am expecting the files to be relatively
small (logos for headers and footers, for example).