Fix selecting all items for Solr based listings.
The provided fix just fixes the issue for tabbedview/table based
listings  in the classical UI.

It works around the fact that the Solr based source does not always
have all contents loaded to apply batching but uses Solr side batching.
This breaks an assumption in the catalog based "select all"
functionality which seems to expect an unbatched list of all brains
being available.

As a workaround we tell the source that we want all results.
The query will be optimized and will only query for the `path` field.