Redirect to notification resource while resolving.
Change the current behavior to always redirect to notification resource
in `@@resolve_oguid` if user has permission to view the resource, even
for foreign notifications. The  previous behavior was to return an
error if the notification was not visible for the current user. The
change adds more convenience in case the link to the notification is
shared between users unintentionally (e.g by forwarding a mail).