Rename English label for "Reference prefix" to "Repository Number":
In technical contexts, this field / concept is still called
"reference number prefix", even though that's incorrect. In any user
facing context, it should be called "Repository number".

This term only refers to the local component of a single
RepositoryFolder. For the full dotted number and the complete reference
number, different terms are used ("rubric" and "reference number",
respectively).