Aggresively expire session to see correct state.
Stale state seems to be visible sometimes in a flaky way which leads to
test failures. This is a testing issue as every action performed by
`accept_forwarding_with_successor` takes place within the same
transaction. I think in combination with the way `dispatch_request`
works in such cases we do not always share the same session.

We aggressively expire everything in the session before querying to
make sure no stale objects stay behind and we get fresh/correct state.