Add support for serializing default_expression via p.supermodel:
This extends the p.supermodel import/export so that it can serialize
and deserialize completely custom attributes that we piggy-back as
plain Python attributes on the IField.

The first of these custom attributes is default_expression, which will
later be exposed via @propertysheets and used to store a TALES expression
that produces a default.