@propertysheets: Add support for setting default_from_member:
This allows to set 'default_from_member' options on fields of a
PropertySheet by specifing a dictionary that describes from which
member property to fetch a default, and optionally, a mapping and/or
fallback value.

This will create a dynamic 'defaultFactory' on the field that will
then fetch and map that member property during runtime, and provide
it as a default value.