Sanitize extracted document_author from mails.
When extracting `document_author` from mails and setting it on D&D
uploaded mails we did bypass input validation of the form fields. We
change how the from header is handled by adding manual sanitation of
extracted values. The field is relatively permissive and only disallows
newlines and carriage returns. We also strip the value before attempting
to lookup an actor so that we could also handle leading/trailing
whitespace should it make its way into the header somehow.