Fix ++widget++ traversal when form contains custom properties:
The ++widget++ traversal for autocomplete searches in the 'responsible'
field cause the entire form to be updated. When the form contained
custom properties widgets, these got updated as well. This caused issues
when the custom property widgets did not contain valid values yet at this
point.

We therefore skip updating custom properties widgets on ++widget++ traversal
requests to avoid this.