Make testserver more sturdy.
The testserver used to expect that between the beginning of a teardown
and the end of the setup no HTTP requests were made.
When requests are made before the testserver is ready again, this
could lead to broken states such as the component registry stacks were
broken, the time freezing implementation broke because of unexpected
time calls while recording, or the ZODB connection were messed up.

In order to make the testserver more sturdy, this change implements a
mechanism causing HTTP requests to be paused when the testserver has
not finished setting up again for the new test.

This mechanism protects the testserver state and makes it hard to
corrupt it.