Truncate filenames when updating favorites.
As the filename column has a length limit we must enforce it by
truncating too long data. This adds another layer of sanitation to the
already existing id normalizer. The normalizer seems to be a bit lenient
when it comes to file extensions potentially allowing very long file
extensions. We add strict sanitation for such cases to prevent errors
while writing to the database.