Fix an issue where we override an existing variable in the transition extender.
We iter through schemas and extend the errors-list. But we assign the errors of the current schema-iteration to the same variable name as the collected errors outside of the for-loop. This will overwrite previously collected errors and will extend the list of errors with itself. This leads to two issues. First, it will only return the errors of the last schema with errors. Second, it will duplicate the errors because it extends the variable with itself.