Fixed copying an object where not all path elements are accessible.
The Copy Service from plone.restapi uses restrictedTraverse to fetch
the copy source. But this only works if every single path element is
accesssible by the current user. But this is not always the
case. Therefore I customized the Copy service and fetch the element in
the same manner the ZCatalog does - fetch the parent unrestricted and
the object itself restricted.

Similar fix was introduced recently for the RelationChoiceFieldDeserializer.