Testserver: provide connection test action for testserverctl.
We want to be able to provide the testserver as docker container.
It is best practice to use health checks so that other containers can
wait until the dependant container is ready.

This connectiontest action can be used as docker / docker-compose
healthcheck. It fails with an exit code until the testserver is up and ready.