Fix storing document in bumblebee during copy-paste.
Copied documents need to get stored in bumblebee once all their
parents have been renamed, as otherwise the url in the storing
job will be outdated by the time the job gets processed. This
only matters for documents with docproperties which will modify
the file, i.e. the document's checksum will not already be stored
in bumblebee.