Fix activity tests.
Activity tests were failing because ActivityTestCase is not an
integration test case and that activities need to resolve actors,
which led to checking whether the KuB feature is enabled. This
would fail because the registry records are not accessible in
unit tests. By modifying the is_kub_contact to first check
the identifier fixes the issue.