Split upgrade in two to mitigate memory issues.
The upgrade to persist some missing default values with nightly jobs
has a relatively high memory consumption as it loads many objects
into memory.

As the upgrade operates on two distinct sets of objects we split the
upgrade into two separate upgrades so that in combination with
intermediate commits the memory can be freed after the first upgrade has
been run.

We introduce a heuristic to detect whether the new upgrade should be run
or not based on checking if the split upgrade with id "20211026132728"
is already installed. For this purpose we also manually backdate our
new upgrade to id "20211026132727" so that it will be picked up and run
immediately before the upgrade we have split into two.