Do not propagate changes of ILifecycle and IClassification anymore.
When modifying a field with a restricted vocabulary on an object
we do not propagate that change to children anymore. Until now
we were propagating to children but only within a depth of 2.
This does not make much sense, but lifting that restriction would
lead to serious performance issues. Also it might not always be
what the user intends as the user might also only change the allowed
values for future content. We therefore decided not to propagate
such changes anymore.