Task delegation does no longer set an unwanted documents-property on the subtask.
The dexterity.utils.createContent method automatically assigns all kwargs to the new content. Currently, the transition extender for delegating tasks forwards the documents-property of the transition transition_params to this function.
This will then create an inpersistent documents-property on the fresh task object. We now get an issue with this behavior because we have had a method called documents on the task. When having a method and a property set on an object, the property will win. So we run into a kind of conflict between property and a callable property. To get rid of this issue (beside of renaming the documents-method in a previous commit), we pop the documents-key from the transition_params instead of just getting it. This will fix the issue for further subtasks.

We decided to not add an upgrade-step to fix the already created documents-properties on the subtasks.