Testserver solr: support backup for remote solr host.
When running the testserver in docker, solr does not run on the same
host because it runs in a separate docker container.

The non-docker implementation directly accesses the solr data on the
disk in order to remove created solr backups (for isolation), since
solr does not provide an endpoint for removing solr backups.

Because we cannot / should not access the solr disk from the
testserver container in docker, we must use unique backup names when
in a multi host environment such as docker.