Fix grouping by resource in digest mail for activities with external resource:
In the digest mail, we used to group notifications by resource. Each resource
was only listed once, with the Plone object's title and a link that directly
resolves it via its OGUID, and then all notifications for that resource
grouped under it.

For external resources this doesn't work any more - we don't have the title
of the external resource, and also no OGUID. Therefore these are listed
in "groups" of one, and their links point to the @@resolve_notification
view in GEVER which will mark the notification as read and redirect.