Dossier journal PDF: Fix updating of existing PDF:
Move the search for an existing document with the
IDossierJournalPDFMarker into the elevated_privileges() block as well.

This fixes a case I noticed when testing locally where, when executed as
nightly job via the zopectl handler, the existing PDF wasn't found unless
an unrestricted search was performed.

There is a test that covers updating of the existing PDF, but it doesn't
catch this special case because the code is not really executed via
zopectl handler.