Add @propertysheet-metaschema endpoint:
This endpoint produces a JSON schema that describes propertysheet definitions.
It will be used by the frontend to render the add/edit forms for property
sheets, and is generated from a zope.schema that will be used to validate the
payload sent to the endpoint (during POST or PATCH).

This is a basic implementation of the endpoint and the meta schema for now,
but eventually it should:
- stricly constrain the property sheets that can be created so that they are
  valid, and can be rendered sensibly in the user interface
- contain all the necessary vocabularies (inline, translated)
- validate default values
- allow validation to produce errors (exceptions) specific enough to map them
  to decent user facing error messages in natural language.