Create titled vocabulary for assignment slots:
This is needed so we can include translatable display names for the
assignment slots we offer to the user, and have those serialized to the
JSON schema.

That way, the frontend can render a (flat) drop down widget for all the
assignment slots, with choices produced by the backend (depending on the
dynamic / customizable document and dossier types).