Add 'id' field to propertysheet metaschema:
This field is not required, and ever actually expected as part of the payload.
For POST, the ID will be the last part of the URL posted to, and for PATCH it
will not be allowed to change the ID of an existing propertysheet.

However, it is part of the schema so it still can be serialized to the JSON
schema, where the frontend can use it for the respective form field in the
add form, and the API endpoint uses it to validate the ID extracted from the
URL.

We also use a custom 'Identifier' schema field for id and name fields:

This field allows identifier-like fields to be constrained to a strict
pattern, that then is also serialized in the JSON schema so that the frontend
can already validate it.