Add 'isidentifier' constraint to id and name fields for propertysheets:
This existing constraint just adds a check that identifiers aren't Python
keywords (the pattern already covers the rest).

This constraint won't (can't) be serialized to the JSON schema, but for now it
exposes the validation that the property sheet definition would do anyway
through the API.

We might also drop this though at some point, since it's most likely
not needed.