Add 'default' field to propertysheet metaschema:
In order to express this field in a valid way in zope.schema, and serialize
it to a JSON schema that content will actually validate against, this field
can't be of a fixed, predetermined datatype.

Therefore we introduce a custom field type MultiTypeField that allows to
set a list of allowed types that the default might accept, and those get
serialized to the JSON, as what effectively works like an "anyOf" constraint
for the field's 'type'.