`@propertysheets` GET: Return same format as POST, instead of JSON schema:
Use the same custom format that POST requests accept as input for GET responses
(and the POST response upon successful creation).

This format, unlike the serialization to a JSON schema, is much more compact,
and more importantly, information-preserving. That way, the `@propertysheets`
endpoint will finally follow the behavior that any response it gives, will
- result in the creation of a equivalent sheet if POSTed
- not result in a meaningful modification if sent as a PATCH payload