Serialize `@propertysheet` errors in a frontend-friendly way:
This overhauls the error handling in the `@propertysheets` APIs in a way that
the client-side errors get serialized in a way where they contain a
deliberately crafted 'translated_message'.

We do this to enable the frontend to just display this one message at the top
of the form, without any further processing of the response or highlighting
of the offending widget.

The message must therefore contain enough context so that the user can
identify where the error is, and what he needs to change to fix it.

This generated `translated_message` is a multi-line string (three, sometimes
four lines), where each line gets progressively more specific / local.

Example:

    The form contains errors:
    Field 3 ('location'):
    Parameter 'title': Value is too long
    (Max: 128 characters, Actual length: 135 characters)

In order to achieve this, we wrap certain validation errors in some custom
higher level exceptions, to which we attach the necessary context info to
build a rich `translated_message` later:

- FieldValidationError (for errors in the `fields` key)
- AssignmentValidationError ((for errors in the `assignments` key)
- SheetValidationError (everything else)

These then get processed by adapters that know how the data attached to these
exceptions is structured, and they extract the necessary context information
from them.

This commit handles POST errors for now, PATCH requires some more changes.
It also adds pre-validation for assignments, which wasn't done on the API
level before.