Testserver: auto-isolate on startup.
The testserver now automatically runs the initial isolation setup on
startup before starting the XMLRPC server.

That means that users and tests no longer have to run a `zodb_setup`
or `isolation` after starting up.
This makes the usage easier and more convenient when the app is doing
requests on startup before the first test.