Reindex 'filename' and 'file_extension' when creating a new document version:
When a new version is saved, the file's type may have changed (e.g. from .pdf
to .docx), and the filename therefore is updated, and needs to be reindexed.

In particular this was the case in a situation where a PDF was reverted (which
creates a new version) to a version that was a Word-Document. In this case
the filename in Solr was outdated, because unlike the catalog, metadata in
Solr doesn't get updated unless the index is explicitly listed.