Add `@raw-dossier-types` endpoint:
This endpoint returns the raw contents of the dossier_types VDEX
vocabulary. They are returned

- in inherent order (even though orderSignificant may be False)
- unfiltered (hidden terms are included)

This is needed in order for the frontend to be able to map dossier types
to colors in a deterministic way that is stable over time.

Because dossier types are mapped to a fixed-length color palette, and
types that have already been mapped to a color once MUST keep that color,
we need to provide the frontend with a list of dossier types where
historical terms keep their position, and new terms only ever are appended
at the end.

This endpoint facilitates this, together with the following rules for
editing the dossier_types.vdex file:

- New terms must only be appended at the end
- Terms are never removed, but set to hidden instead (registry).
  They will then be omitted from the regular vocabularies / sources,
  which are elephant vocabs, but not this endpoint.
- Renames are ok (if the type is supposed to keep its color)