Default value acquisition: Skip intermediate objects missing attribute:
When acuiring default values, the acquire_field_value() may encounter
objects in the chain that do claim to provide the field, but are missing
the actual attribute because it never got persisted properly.

Because we now use get_persisted_value_for_field() to fetch the field value,
we need to account for the AttributeError it may raise (unlike the
field.get() that was used before).