Provide `@reference-number` expansion and view
Provide an expandable view `@reference-number` that returns the
formatted reference number, a sortable reference number and the
unformatted reference number parts.

This helps client applications that need to use the raw reference number
or parts of the raw reference.

I did not find an interface for which to register a single service.
Instead i found that the `IReferenceNumber` adapter is registered for
`ISiteRoot` and `IDexterityContent`. Thus the new service and adapter
mirror the adapter registrations of `IReferenceNumber`. The same service
and adapter are registered twice for different interfaces.