Provide an example docker-compose file for testserver.
The purpose minimal docker-compose file for the testserver is to act
as an example for copying to other projects. It is simplified to a
minimal workable version and contains a commented example for mounting
a fixture module as volume, because that is common for third party
projects using the testserver.