NotificationDispatcher: Only return failed notification as not_dispatched:
The intent here was to return any failed notifications in an {'errors': [..]}
list. However, for each notification that failed to be dispatched *all* the
notifications were added to the error list.

This hasn't surfaced yet because it wasn't covered by tests, and the return
value wasn't really used until now.