SolrReporterView: Add support for mapping Excel columns to TabbedView columns:
This is necessary because otherwise in the old UI columns get dropped from
the Excel if the reporter view can't map it to a known tabbedview column.

Just defaulting to the 'sort_index' on the ListingFields gets us 90% there.
Most TabbedView columns are named the same as the sort_index.

One exception is the 'reference' field: Its sort_index is 'sortable_reference',
but the TabbedView column is actually called 'reference'. For these cases,
a 'tabbedview_column' key can be specified in the reporter's column_settings.

Fixes one of the two failing tests.