Support aliasing of columns in SolrReporterView:
This is needed because the frontend names some columns in a way that
deviates from the name of the ListingField that we would want to use for the
Excel report (for example, 'responsible_fullname' instead of 'responsible').

And to keep logic out of the frontend, we want it to just pass its list of
user visible columns 1:1 to the respective reporter view.

While we could get around this for some cases by carefully choosing the
column names we pick in the reporter's column_settings, introducing support
for aliasing columns if necessary allows us to keep column definitions clean,
and have a flexible mechanism to deal with such cases in the future.