Remove flaky test.
The test to copy a document from the second batch from a workspace
does not make sense, as there is no such thing as a document from
the second batch (the search is not sorted). So this test was
flaky and I don't see any way to make a meaningful test for this.