Revert partially "Fix translated review state for meeting content:"
This reverts commit ec89e2dccdc4facd3d8f234b202948612158d56b.

'get_state' will do an object lookup which decreases the performance of a listing. We can't use this method anymore for proposals.
This implementation seems not to be an issue for proposals in general but for submitted proposals. We do not list submitted proposals anywhere, so we just remove it.