Replace WorkspaceMember vocabularies with WorkspaceContentMember.
The ActualWorkspaceMember source, vocabulary and endpoint did not
make much sense, as they basically returned the list of participants
on a workspace. This was done by walking up the acquisition chain,
but starting at the workspace and stopping at the portal, so basically
just gathering the participants of the workspace (and workspace root,
but that does not have any participants). Instead these vocabularies
were always meant to be context dependent and walk up the acquisition
chain. They are used for ToDo and WorkspaceMeeting responsibles as
well as for sharing content from a workspace content. For ToDo and
WorkspaceMeeting responsibles, things worked as expected as they
are only allowed in the Workspace and not inside WorkspaceFolders.
The sharing itself was broken on other contents as the vocabulary
basically allowed choosing all workspace participants, regardless
of whether inheritance is blocked for example on the WorkspaceFolder.
This used to work but got broken at some point, probably because
the naming seemed to indicate the intent was to return the
Workspace members. I have therefore also renamed the vocabulary,
source and endpoint.