Introduce language specific text fields
Enable language detection and index text fields with language specific
analyzers. In general we perform stemming to avoid using wildcards in
queries and to be able to switch to the eDisMax parser.

Fields are still indexed using the canonical name. The langid update
processor then detects the language and appends the language suffix to
field name. However querying needs to be done against the field with the
language suffix.