Preserve original token when stemming
Because stemming is not applied to wildcard queries, they do not work as
expected. E.g. the term "Verträge" is indexed as "vertrag" because of
normalization and stemming. Searching for "verträge*" will not match,
because no stemming is applied to a wildcard query. To work around this,
we also index the unstemmed token. Further queries matching the exact
term should get a better score.