Retrieve comment from event for DocumenVersionCreatedActivity.
The comment is already stored on the corresponding event
(ObjectCheckedInEvent), so we can avoid retrieving the latest
version of the document in the activity by simply getting the
comment from the event. This is more efficient and fixes issues
with testing, as retrieving a version makes a transaction savepoint
which is not supported by the testserver.