Fix user lookup by email for inbound mail:
To avoid fuzzy matches when looking up the user by email address for inbound
mails, the `pas_search` call was extended with `exact_match=True` to
accomplish this.

However, it turns out that the `ActiveDirectoryMultiPlugin` refuses to do
searches with `exact_match=True`, unless the query criteria contain the user's
id or login. For the `LDAPMultiPlugin` this works fine, but the AD plugin
raises a ValueError that later gets supressed, and so no matches are returned.

We therefore change the strategy to let `pas_search` return fuzzy matches
again, and filter them post-search to only pick case-insensitive, but otherwise
exact matches.