Make enumerateUsers() return all users if no search criteria are given:
This is not explicitly required as per the IUserEnumerationPlugin interface,
but it's what the LDAP and AD plugins from P.LDAPMultiPlugins do.

It's used by the Plone Control Panel to list all users, but not, if memory
serves me right, by the `@users` endpoint in plone.restapi. That one will only
fully list all users for plugins that implement IUserIntrospection from
PlonePAS.