Support `exact_match` kwarg for enumerateUsers() and enumerateGroups():
With exact_match=True, user/group records are looked up by ID or login
(currently still case-sensitive). One of those criteria is therefore required,
otherwise a ValueError is raised (same behavior as the LDAP/AD plugins).

In this case, Postgres could use the already existing BTree indexes on the
User / Group tables' primary key columns. It seems to decide that a full table
scan is faster though in situations with not that many rows.

For exact_match=False, a case-insensitive substring search is performed using
ILIKE '%searchterm%'. This currently always requires a full table scan. If
raw query performance (before caching) needs to be optimized for this case,
this could be done with a Trigram GIN index, using the builtin pg_trgm
extension.

The `WHERE active = true` clause that is part of almost every query can't use
an index, because those columns don't have one yet. But at least in the case
of exact_match=True it doesn't matter, because the filtering by ID is already
extremely selective (eliminates all rows but one).

It might matter for listing all (active) users or groups, but those are far
less important.