Filter groups with non-ASCII names from getGroupsForPrincipal():
This is necessary to make Plone function correctly when groups with non-ASCII
names are present. Those groups still can't be used, but their mere presence
at least does not cause problems if we filter them out.

This carries over the fix we did in 4teamwork/Products.LDAPMultiPlugins#3
to the OGDS PAS plugin.