Don't return any results when presented with unknown search criteria:
When either enumerateUsers() or enumerateGroups() are called with keyword
arguments that don't map to an SQL column that we can actually search, we
should return zero results.

The rationale behind this is that the 'sharing' browser view, when called
with a 'search_term' to look up principals to add, dispatches several
(separate) searches for users and groups with "well-known" attribute names
like "name", "fullname" and "email" (for users) or "title" and "description"
(for groups).

It then merges all those search results together, and returns them. With the
existing implementation, where unknown search criteria just got skipped, this
resulted in all users and groups always being returned, because at least one
of those criteria was unknown to us, and lead to all items being returned
(because no other search criteria were present).