Make max dossier depth restriction less strict:
Don't enforce max dossier depth if a structure of same nesting level as the
new one already exists.

For example: If a dossier already contains a subdossier with a sub-sub-dossier
that already exceeds the max depth by 2, then it will be allowed to create more
content on that level that exceeds the depth by up to 2 levels (but not more).