Support searching by `fullname` in enumerateUsers():
This is needed for the sharing view (and therefore `@sharing` endpoint) to be
able to search for principals via a search term.

Because we don't have a real `fullname` column in OGDS yet, we need to emulate
it in order for it to be queried. By adding a SQLAlchemy column expression
that concatenates firstname and lastname, we make 'fullname' queryable.

Incidentally, this also leads to the 'fullname' column epxression being
selected when getting properties for a user, so we can eliminate the special
handling we had for fullname there.