Enable sharing view to search users by login:
The sharing view, when queried for principals with a search_term, never
directly queries pas_search.searchUsers() for 'id' or 'login'. Instead, it
queries ['name', 'fullname', 'email'] for the search term, and merges the
results.

pas_search.searchUsers() then, when queried for 'name', extends the search
criteria to also query 'login' for the same search term. Because we don't
have an actual 'name' column or property, that query never matches anything
and only users' fullname and email were searchable.

By just dropping the 'name' criterion (when login is also given), we turn
this into a query that just searches the user's 'userid' column, therefore
making the sharing view able to also filter by userid.