Add `display_name` column to OGDS user model:
The LDAP/AD attributes 'displayName' or 'cn' are sometimes
mapped to the 'fullname' property. This is especially the case for customers
that actively manage the display names of their users, e.g. to also contain
the user's department or their role.

If it exists, we store it in OGDS in the display_name column. This is therefore
a customer-controlled display name, and different from the fullname we build
in the User.fullname() method by concatenating first and last name.