Fix how we access SA RowProxy results in OGDS auth plugin:
When using SQLAlchemy Core queries instead of the ORM, the returned results
aren't ORM mapped instances, but RowProxy objects that look and act a lot
like tuples:

(u'Peter', u'Archivar', u'peter.archivar', None, u'info@4teamwork.ch')

Usually, the returned values are exactly in the order of the columns in the
SELECT statement, which is why it's tempting to jsut consider them a tuple
in the same order as the selected columns, and unpack the values that way.

However, if a column appears more than once in a SELECT statement, the
(identical) return value will not be repeated in the RowProxy result, which
is then shorter than the list of selected columns, throwing the whole order
off.

Instead, the RowProxy object also supports dictionary access by using the
column name as the key, which makes access to the column results unambiguous
and lets us avoid that whole ordered sequence matching dance.