Add `organization` column to OGDS user model:
This column is intended to track a user's organization in a simpler way than
the existing directorate and department columns.

In order to be populated, it will need to be mapped in the LDAP schema to an
AD/LDAP attribute first, with a public_name of 'organization'. Then it will be
synced to OGDS.