Add support for omitting app prefix in BaseDocPropertyProvider.get_properties():
When `with_app_prefix=False` is specified as a keyword argument, the properties
are still prefixed with a possible local prefix and default prefix, but not
with the app namespace ('ogg').

This allows for easier reuse of DocPropertyProviders to build composite
providers that delegate to other providers, and include their results (which
may also contain already nested namespaces).