Include KuB organization phone number in membership docproperties:
When serializing a KuB membership to DocProperties, this will also include the
organization's phone number in 'ogg.[scope].organization.phone.number'
(where [scope] may currently be 'sender' or 'recipient').

A side effect of this is that when serializing an 'organization' entity, its
phone number is now exposed twice in DocProperties: Once in
'ogg.[scope].organization.phone.number' and once in
'ogg.[scope.]phone.number'.