Add `@prepare-copy-dossier-to-workspace` endpoint:
This API endpoint is used to prepare copying a dossier to a linked workspace.

It operates in two modes: Validation, or preparing an empty folder structure on
the teamraum side to upload documents to.

In validation mode, it checks that the given structure (subdossier(s) and
documents) satisfies all the constraints to be copied to a workspace. If not,
a detailed list of errors by object is returned.

In "prepare structure" mode, it will mirror the structure of the given
subdossier to teamraum by creating empty folders in the workspace, and
returning a list of documents, including their target folders that the client
should then upload them to.

The second stage ("prepare structure") does not perform validation again. It
assumes that the client validated the structure beforehand, and either
everything is ok, or the user selected "copy anyway", in which case some
individual document uploads will fail.