Track if a document has pending changes
We want to know if a document was modified since the last checkout and
if there are changes that haven't been checked in yet.
Unfortunately we can not determine this directly as Word documents are
always modified on checkout and checkin because of DocProperties
updates. Therefore we add a marker when a new document version gets
uploaded.